+++ library/LibUtilities/BasicUtils/MeshPartition.cpp
@@ -1140,6 +1140,23 @@ namespace Nektar
             }
         }
 
+        /**
+         * @brief Partition the graph.
+         *
+         * This routine partitions the graph @p pGraph into @p nParts, producing
+         * subgraphs that are populated in @p pLocalPartition. If the @p
+         * overlapping option is set (which is used for post-processing
+         * purposes), the resulting partitions are extended to cover
+         * neighbouring elements by additional vertex on the dual graph, which
+         * produces overlapping partitions (i.e. the intersection of two
+         * connected partitions is non-empty).
+         *
+         * @param pGraph           Graph to be partitioned.
+         * @param nParts           Number of partitions.
+         * @param pLocalPartition  Vector of sub-graphs representing each
+         *                         partition.
+         * @param overlapping      True if resulting partitions should overlap.
+         */
         void MeshPartition::PartitionGraph(BoostSubGraph& pGraph,
                                            int nParts,
                                            std::vector<BoostSubGraph>& pLocalPartition,
@@ -1257,6 +1274,8 @@ namespace Nektar
                 boost::add_vertex(i, pLocalPartition[part[i]]);
             }
 
+            // If the overlapping option is set (for post-processing purposes),
+            // add vertices that correspond to the neighbouring elements.
             if(overlapping)
             {
                 //